refactor(chat) : comment -> chat 을 이름 변경 및 rds 나 sql 로 채팅 저장을 안할거기 때문에 엔티티제거 (s3 나 DynamoDb 에 저장할 예정)
