+++ src/main/java/com/gamemetricbackend/domain/comment/dto/ChatRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class CreateCommentDto {
+public class ChatRequestDto {
     private String comment;
     private Long broadCastId;
 }
