--- src/main/java/com/gamemetricbackend/domain/comment/repository/CommentRepositoryQuery.java
@@ -1,4 +1,0 @@
-package com.gamemetricbackend.domain.comment.repository;
-
-public interface CommentRepositoryQuery {
-}
