+++ src/main/java/com/gamemetricbackend/domain/comment/service/ChatService.java
@@ -1,0 +1,6 @@
+package com.gamemetricbackend.domain.comment.service;
+
+public interface ChatService {
+
+    Boolean CreateChat(Long id, Long broadcastId,String comment);
+}
