--- src/main/java/com/gamemetricbackend/domain/comment/service/CommentService.java
@@ -1,6 +1,0 @@
-package com.gamemetricbackend.domain.comment.service;
-
-public interface CommentService {
-
-    void CreateComment(Long id, Long broadcastId,String comment);
-}
